Move asset publishing to an external script.
