+++ _config.yml
@@ -69,8 +69,13 @@ kramdown:
     coderay_bold_every: 10
     coderay_css: class
 
-include: [".htaccess"]
-exclude: ["lib", "config.rb", "Capfile", "config", "log", "Rakefile", "Rakefile.rb", "tmp", "less", "*.sublime-project", "*.sublime-workspace", "test", "spec", "Gruntfile.js", "package.json", "node_modules"]
+include:
+- .htaccess
+exclude:
+- '*.sh'
+- README.md
+- LICENSE
+- Gemfile*
 
 assets:
   sources:
