+++ _fog.yml.example
@@ -2,10 +2,3 @@ account:
   username: smashwilson
   api_key: abcdefabcdef
 container: azurefire
-files:
-- assets/js/scripts.min.js
-directories:
-- assets/css
-- assets/fonts
-- assets/img
-- images
